+++ src/main.ts
@@ -14,7 +14,7 @@ const ListPanel = () => import('./components/list/ListPanel.vue')
 const CustomPanel = () => import('./components/CustomPanel.vue')
 const VoiceSearchPanel = () =>
   import('./components/search/VoiceSearchPanel.vue')
-const BgmPanel = () => import('./components/BgmPanel.vue')
+const BgmPanel = () => import('./components/bgm/BgmPanel.vue')
 
 // create a new router instance
 const router = createRouter({